+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2] - 2025-04-17
+
+- Add "Bump Release Number" function.
+- Expand macro when used in `Version` directive.
+- Add Epoch number to changelog.
+
 ## [0.1] - 2025-03-10
 
 - Initial release